requirement.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
pycodestyle==2.12.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
